refactor: split manual rebind module (#4157)
