+++ scripts/giant_file_registry.toml
@@ -572,17 +572,6 @@ deadline = "2026-10-31"
 decompose_issue = "#3405"
 
 [[entry]]
-# Crossed the 1000-prod-LoC giant threshold in #4091 r3/r4 when manual rebind
-# gained ClaudeTui transcript adoption (candidate probing, EOF-forced start
-# offsets for adopted transcripts, runtime-binding re-registration).
-# Decomposition should lift the transcript-adoption coordinate decisions and
-# binding re-registration into narrower siblings next to manual_rebind_output_path.rs.
-file = "src/services/discord/recovery_engine/manual_rebind.rs"
-owner = "discord-relay"
-deadline = "2026-10-31"
-decompose_issue = "#4157"
-
-[[entry]]
 # Crossed 1000-prod-LoC at #4150 r4 (two-signal anchor + delivered-elsewhere
 # frontier probe with same-coordinate proof + author detection, 9-row table).
 file = "src/services/discord/tmux_placeholder_suppression.rs"
